chore: consolidate controller, host and runtime packages
