+++ cmd/server.go
@@ -32,8 +32,8 @@ import (
 	"github.com/flanksource/incident-commander/mcp"
 	"github.com/flanksource/incident-commander/metrics"
 	"github.com/flanksource/incident-commander/notification"
+	"github.com/flanksource/incident-commander/plugin/machinery"
 	pluginregistry "github.com/flanksource/incident-commander/plugin/registry"
-	"github.com/flanksource/incident-commander/plugin/supervisor"
 	echov4 "github.com/labstack/echo/v4"
 
 	// register event handlers & echo routers
@@ -158,7 +158,7 @@ func launchKopper(ctx context.Context) {
 		shutdown.ShutdownAndExit(1, fmt.Sprintf("Unable to create controller for Team: %v", err))
 	}
 
-	supervisor.Wire(ctx)
+	machinery.Wire(ctx)
 	if _, err := kopper.SetupReconciler(ctx, mgr,
 		pluginregistry.PersistPluginFromCRD,
 		pluginregistry.DeletePlugin,
