+++ plugin/controller/connection.go
@@ -1,4 +1,4 @@
-package host
+package controller
 
 import (
 	"encoding/json"
