+++ plugin/controller/interceptor.go
@@ -1,4 +1,4 @@
-package host
+package controller
 
 import (
 	"context"
