+++ plugin/controller/invoke.go
@@ -1,4 +1,4 @@
-package runtime
+package controller
 
 import (
 	"context"
