--- plugin/host/suite_test.go
@@ -1,13 +1,0 @@
-package host
-
-import (
-	"testing"
-
-	ginkgo "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestHost(t *testing.T) {
-	RegisterFailHandler(ginkgo.Fail)
-	ginkgo.RunSpecs(t, "Host")
-}
